Finish rename
