+++ packages/app/src/features/actions/flavours/claim-rewards/logic/claimRewardsAction.test.ts
@@ -3,7 +3,7 @@ import { aaveDataLayerQueryKey } from '@/domain/market-info/aave-data-layer/quer
 import { getBalancesQueryKeyPrefix } from '@/domain/wallet/getBalancesQueryKeyPrefix'
 import { getMockToken, testAddresses } from '@/test/integration/constants'
 import { handlers } from '@/test/integration/mockTransport'
-import { setupUseActionRenderer } from '@/test/integration/setupUseActionRenderer'
+import { setupUseContractActionRenderer } from '@/test/integration/setupUseContractActionRenderer'
 import { waitFor } from '@testing-library/react'
 import { mainnet } from 'viem/chains'
 import { describe, test } from 'vitest'
@@ -15,7 +15,7 @@ const incentiveControllerAddress = testAddresses.token
 const assets = [testAddresses.token2]
 const token = getMockToken()
 
-const hookRenderer = setupUseActionRenderer({
+const hookRenderer = setupUseContractActionRenderer({
   account,
   handlers: [handlers.chainIdCall({ chainId }), handlers.balanceCall({ balance: 0n, address: account })],
   args: { action: { type: 'claimRewards', incentiveControllerAddress, assets, token }, enabled: true },
