+++ packages/app/src/features/actions/flavours/set-use-as-collateral/logic/setUseAsCollateralAction.test.ts
@@ -3,7 +3,7 @@ import { lendingPoolAddress } from '@/config/contracts-generated'
 import { TokenSymbol } from '@/domain/types/TokenSymbol'
 import { getMockToken, testAddresses } from '@/test/integration/constants'
 import { handlers } from '@/test/integration/mockTransport'
-import { setupUseActionRenderer } from '@/test/integration/setupUseActionRenderer'
+import { setupUseContractActionRenderer } from '@/test/integration/setupUseContractActionRenderer'
 import { waitFor } from '@testing-library/react'
 import { mainnet } from 'viem/chains'
 import { describe, test } from 'vitest'
@@ -13,7 +13,7 @@ const collateral = getMockToken({ symbol: TokenSymbol('TEST') })
 const account = testAddresses.alice
 const chainId = mainnet.id
 
-const hookRenderer = setupUseActionRenderer({
+const hookRenderer = setupUseContractActionRenderer({
   account,
   handlers: [handlers.chainIdCall({ chainId }), handlers.balanceCall({ balance: 0n, address: account })],
   args: { action: { type: 'setUseAsCollateral', token: collateral, useAsCollateral: true }, enabled: true },
